+++ src/search/writer/writers/var_writer.rs
@@ -7,7 +7,7 @@ pub struct VarWriter<const C: usize> {
 }
 
 impl<const C: usize> VarWriter<C> {
-    pub fn new(_length: usize, _constant_cap: u8) -> Self {
+    pub fn new(_length: usize, _constant_cap: u128) -> Self {
         Self {
             next_write: 0,
         }
